+++ src/pages/ItemsPage/ItemsPage.jsx
@@ -1,8 +1,9 @@
 import "./ItemsPage.css";
 import { usePageSizeByBreakPoint } from "../../hooks/usePageSizeByBreakPoint";
-import BestItemsSection from "../../components/BestItemsSection";
-import CurrentItemsSection from "../../components/CurrentItemsSection";
+
 import Nav from "../../components/layout/Nav/Nav";
+import BestItemsSection from "./sections/BestItemsSection";
+import CurrentItemsSection from "./sections/CurrentItemsSection";
 
 const ItemsPage = () => {
   const { pageSizeList } = usePageSizeByBreakPoint();
